Convert module to adapter-algorithm pattern
