+++ algorithms/twoAxisGimbalAxisToMotorAngles/CMakeLists.txt
@@ -4,6 +4,7 @@ if(_is_enabled)
   xmera_add_swig_module("${module}")
   target_sources("${module}" PRIVATE
     twoAxisGimbalAxisToMotorAngles.cpp
+    twoAxisGimbalAxisToMotorAnglesAlgorithm.cpp
   )
 
   target_include_directories("${module}" PUBLIC
